==> workbench/1_data_extraction.ipynb <==
## `Flight Delay Prediction`
### **Part 1:** Data Extraction

Ali Bahrami


# import libraries
import numpy as np
import pandas as pd
import psycopg2
import config as cfg


## Extract the Data

# Connect to the postgresql database.
# The database credentials are stored separately in config.py to avoid sensitive data from being uploaded to github
user = cfg.database['user']
password = cfg.database['password']
host = cfg.database['host']
port = cfg.database['port']
database = cfg.database['database']

conn = psycopg2.connect(user = user,
                        password = password,
                        host = host,
                        port = port,
                        database = database)

query = "select * from flights ORDER BY RANDOM() LIMIT 200000"

df_flights = pd.read_sql(sql=query ,con=conn)
df_flights.shape

query = "select * from passengers ORDER BY RANDOM() LIMIT 200000"

df_passengers = pd.read_sql(sql=query ,con=conn)
df_passengers.shape

query = "select * from fuel_comsumption"

df_fuel = pd.read_sql(sql=query ,con=conn)
df_fuel.shape

query = "select * from flights_test"

df_flights_test = pd.read_sql(sql=query ,con=conn)
df_flights_test.shape

# Export the files into csv
df_flights.to_csv("data/raw/flights_raw.csv", index=False)
df_flights_test.to_csv("data/flights_test.csv", index=False)
df_passengers.to_csv("data/raw/passengers_raw.csv", index=False)
df_fuel.to_csv("data/raw/fuel_raw.csv", index=False)

# Close the cursor and connection to so the server can allocate bandwidth to other requests
conn.close()
