==> workbench/2_data_preprocessing.ipynb <==
## `Flight Delay Prediction`
### **Part 2:** Data Preprocessing for EDA

-- This is just a mass cleanup of the data. The data will be polished and cleaned further in Part 3: EDA.

Ali Bahrami

import numpy as np
import pandas as pd

# Read from extracted data
df_flights_raw = pd.read_csv("data/raw/flights_raw.csv")
df_passengers_raw = pd.read_csv("data/raw/passengers_raw.csv")
df_fuel_raw = pd.read_csv("data/raw/fuel_raw.csv")

# Take copy to measure data loss after clean-up
df_flights = df_flights_raw.copy()
df_passengers = df_passengers_raw.copy()
df_fuel = df_fuel_raw.copy()

### Missing Values Treatment

df_flights.head()

Drop cancelled and diverted flights, given that the focus of the project is only on flights delays


df_flights = df_flights[(df_flights['cancelled'] != 1)&(df_flights['diverted']!=1)]

Drop unnecessary columns:
- No need for mkt_carrier data as there is operating carrier
- cancellation codes are not needed.
- flight numbers are not needed.
- airport ids are not needed.
- other columns with more than 90% of data missing 

cols = ['cancellation_code' , 'mkt_carrier','op_unique_carrier', 'branded_code_share', 'mkt_carrier_fl_num', 'op_carrier_fl_num', 'diverted','first_dep_time', 'origin_airport_id', 'dest_airport_id', 'flights','total_add_gtime', 'longest_add_gtime', 'no_name']
df_flights = df_flights.drop(columns=cols)

df_flights.isnull().sum()

The columns with reasons of delay (carrier delay, weather delay, etc.) all seem to similar amounts of missing data: 158311

which is more than 50% of the entire dataset. 

I will fill them with zero for now, but will drop them during the modeling phase.

# Fill na values in appropriate columns with zero.
# i.e. security_delay either has values in minutes or na values which would be zero

columns_fill = ['carrier_delay', 'weather_delay', 'nas_delay', 'security_delay', 'late_aircraft_delay']
df_flights.update(df_flights[columns_fill].fillna(0))

# Drop any sporadic nan values (i.e. ghost flights with no air time) (207 values)
df_flights.dropna(axis=0, inplace=True)

df_flights.isnull().sum()

## Preliminary Data Manipulation and Feature Engineering

# The airlines names are given as abbreviations only, import the full name from the files.
airlines = pd.read_csv("data/airlines.csv")
abbr_airlines = airlines.set_index('IATA_CODE')['AIRLINE'].to_dict()

# Add the full airline names to the dataframe
df_flights['carrier'] = df_flights['mkt_unique_carrier'].replace(abbr_airlines)

df_flights['carrier'].unique()

Day of Week GLOSSARY

0 = Monday

1 = Tuesday

2 = Wednesday

3 = Thursday

4 = Friday

5 = Saturday

6 = Sunday

# Add separate columns for year, month, day of month, and day of week
df_flights['fl_date'] = pd.to_datetime(df_flights['fl_date'],format='%Y-%m-%d')
df_flights['month'] = pd.DatetimeIndex(df_flights['fl_date']).month
df_flights['day_of_week'] = pd.DatetimeIndex(df_flights['fl_date']).dayofweek

Divide the flight into short, medium, and long haul flights based on air time


def haul(df, col):
    '''Add a column to determine flight haul (short, medium, long) based on air_time'''
    length=[]
    for i in df[col]:
        if i < (3*60): # up to 3 hours
            length.append('short')
        elif (i >= (3*60)) and (i < (6*60)): # 3-6 hours
            length.append('medium')
        elif i >= (6*60):
            length.append('long') # 6+ hours
    df_flights['haul_length'] = length

haul(df_flights, 'air_time')

# Add flight route
df_flights['route'] = df_flights['origin'] + "-" + df_flights['dest']

### Binary Classification
Classify the flights in a new column 'flight_delayed' into:

delayed = 1

not delayed = 0

status = []

for value in df_flights['arr_delay']:
    if value <= 0:
        status.append(0)
    else:
        status.append(1)
df_flights['flight_delayed'] = status


df_flights.head() 

## Passengers Dataset
No missing values in Passengers

df_passengers.isnull().sum()

## Fuel Consumption Dataset
Drop rows with missing data in fuel consumption (25 values in 3035 rows)

df_fuel = df_fuel.dropna()

## Data Loss

# Data Loss

# Flights
flights_raw_len = len(df_flights_raw)
flights_clean_len = len(df_flights)
flights_data_loss = round(((flights_raw_len - flights_clean_len)/flights_raw_len  * 100), 2)
print(f'Flights data lost in percent: {flights_data_loss}%')

# Passengers
passengers_raw_len = len(df_passengers_raw)
passengers_clean_len = len(df_passengers)
passengers_data_loss = round(((passengers_raw_len - passengers_clean_len)/passengers_raw_len  * 100), 2)
print(f'Passengers data lost in percent: {passengers_data_loss}%')

# Fuel Consumption
fuel_raw_len = len(df_fuel_raw)
fuel_clean_len = len(df_fuel)
fuel_data_loss = round(((fuel_raw_len - fuel_clean_len)/fuel_raw_len  * 100), 2)
print(f'Fuel data lost in percent: {fuel_data_loss}%')

## Export Cleaned Data

df_flights.to_csv("data/flights.csv", index=False)
df_passengers.to_csv("data/passengers.csv", index=False)
df_fuel.to_csv("data/fuel.csv", index=False)
