==> workbench/5_feature_selection_test.ipynb <==
## `Flight Delay Prediction`
### **Part 2:** Data Cleaning and Feature Selection

Ali Bahrami

import numpy as np
import pandas as pd
import matplotlib as mpl
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.graph_objects as go


from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import StandardScaler

%matplotlib inline


# Read from extracted data
df = pd.read_csv("flights_test.csv")

# Perform all the operations conducted on training data to the evaluation data.
# To save time, df_flights in this notebook is taken from flights_test
df_flights = df.copy()
df_flights.head()

# The airlines names are given as abbreviations only, import the full name from the files.
airlines = pd.read_csv("airlines.csv")
abbr_airlines = airlines.set_index('IATA_CODE')['AIRLINE'].to_dict()

# Add the full airline names to the dataframe
df_flights['carrier'] = df_flights['mkt_unique_carrier'].replace(abbr_airlines)


# Add separate columns for year, month, day of month, and day of week
df_flights['fl_date'] = pd.to_datetime(df_flights['fl_date'],format='%Y-%m-%d')
df_flights['month'] = pd.DatetimeIndex(df_flights['fl_date']).month
df_flights['day_of_week'] = pd.DatetimeIndex(df_flights['fl_date']).dayofweek





def bin_hours(time):
    # Bin the time from HHMM input format to 24 hrs where 23=11pm and 0=12am
    if type(time) == 'numpy.ndarray':
        time = time.astype(int)
    time = int(time)
    
    t = time // 100
    
    if t == 0:
        return 0
    return t

# Bin the departure time into hours
df_flights.loc[:,'crs_dep_time'] = df_flights['crs_dep_time'].apply(bin_hours)
df_flights.loc[:,'crs_arr_time'] = df_flights['crs_arr_time'].apply(bin_hours)



mean_hist_dep_delay = pd.read_csv("mean_hist_dep_delay.csv")
mean_hist_arr_delay = pd.read_csv("mean_hist_arr_delay.csv")

merged = pd.merge(df_flights, mean_hist_dep_delay, on=['crs_dep_time'], how='left')
df_flights = pd.merge(merged, mean_hist_arr_delay, on=['crs_arr_time'], how='left')

df_flights.head()

## We had included the air_time by error and to complete the assignment, I am gonna input the mean here, but this needs to be fixed
df_flights.insert (6, "air_time", 107.6973116856836)

df_flights = df_flights[['origin', 'dest', 'crs_dep_time', 'crs_arr_time',
       'crs_elapsed_time', 'air_time', 'distance', 'carrier', 'month',
       'day_of_week', 'mean_hist_dep_delay', 'mean_hist_arr_delay']]

df_flights.head()

## Label Encode

# Origin
le_origin = LabelEncoder()
le_origin.fit(df_flights.origin)
df_flights.loc[:,"origin"] = le_origin.transform(df_flights.origin)

# Dest
le_dest = LabelEncoder()
le_dest.fit(df_flights.dest)
df_flights.loc[:,"dest"] = le_dest.transform(df_flights.dest)

# Carrier
le_carrier = LabelEncoder()
le_carrier.fit(df_flights.carrier)
df_flights.loc[:,"carrier"] = le_carrier.transform(df_flights.carrier)

df_flights.dtypes

df_flights.head()

df_flights.shape

## Scale the data and prepare it for modeling

df_model = df_flights.copy()

X = df_model

scaler = StandardScaler()
scaler.fit(X)
scaled_df = scaler.fit_transform(X)

df_flights.columns

flights_scaled = pd.DataFrame(data = scaled_df,
              columns = ['origin', 'dest', 'crs_dep_time', 'crs_arr_time',
       'crs_elapsed_time', 'air_time', 'distance', 'carrier', 'month',
       'day_of_week', 'mean_hist_dep_delay', 'mean_hist_arr_delay'])

flights_scaled

# This is the scaled test data that we will use to predict the model on.
flights_scaled.to_csv("data/test_eval.csv", index=False)
